# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VehiGo API docs",
        "description": "This is an API documentation, generated by L5 Swagger",
        "contact": {
            "email": "office@atelierulmeu.ro"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get user notification by id",
                "description": "Returns the notification detailed data",
                "operationId": "9cd1492cfb2e03e7ca29dc21ebbce9f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of notification",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification response by notification id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notification"
                ],
                "summary": "Update the current notification",
                "description": "Updates the user notification by id",
                "operationId": "151388cc957a8aa19bc3aba10b7fe9ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of notification",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Notification fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated notification response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notification"
                ],
                "summary": "Delete the user notification",
                "description": "Deletes the user notification by id",
                "operationId": "c45b0ae13c892255f8acb67fb481ac8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of notification",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete notification response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteNotificationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
        "/notifications": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get user notifications",
                "description": "Returns the user related notifications",
                "operationId": "4c4c62ede38b4d58cdad5ae89a516a8a",
                "responses": {
                    "200": {
                        "description": "The user notifications response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserNotificationsResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Create new user notification",
                "description": "Creates a new user related notification",
                "operationId": "dfe9686351c9782afdd01984f1e4ad15",
                "requestBody": {
                    "description": "Notification fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user data",
                "description": "Returns the current user data.",
                "operationId": "d8dabdeba82df94aeebbc9bb35076a1f",
                "responses": {
                    "200": {
                        "description": "The current user data response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user data",
                "description": "Returns the updated user data.",
                "operationId": "18519f226287bd2712bc626fbb2d1880",
                "requestBody": {
                    "description": "User fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated user data response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user data",
                "description": "Returns a success meassge from delete user process..",
                "operationId": "d2c6d1447ce53e08364450f3f8d7d651",
                "responses": {
                    "200": {
                        "description": "The updated user data response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteUserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "NotificationResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "$ref": "#/components/schemas/Notification"
                    }
                },
                "type": "object"
            },
            "DeleteNotificationResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserNotificationsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notification"
                        }
                    }
                },
                "type": "object"
            },
            "NotificationBody": {
                "required": [
                    "car_id",
                    "type",
                    "title",
                    "alert_date"
                ],
                "properties": {
                    "car_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "ITP"
                    },
                    "title": {
                        "type": "string",
                        "example": "Audi ITP"
                    },
                    "message": {
                        "type": "string",
                        "example": "My message"
                    },
                    "alert_date": {
                        "type": "string",
                        "example": "2023-06-05 10:00:00"
                    },
                    "expiration_date": {
                        "type": "string",
                        "example": "2023-06-06 10:00:00"
                    },
                    "meta_data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateUserBody": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "George"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Tepes"
                    },
                    "username": {
                        "type": "string",
                        "example": "g_tepes"
                    },
                    "email": {
                        "type": "string",
                        "example": "tepes@vlad.ro"
                    }
                },
                "type": "object"
            },
            "DeleteUserResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
...
            "Notification": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "ITP"
                    },
                    "title": {
                        "type": "string",
                        "example": "Audi ITP"
                    },
                    "message": {
                        "type": "string",
                        "example": "My message"
                    },
                    "alert_date": {
                        "type": "string",
                        "example": "2023-06-05 10:00:00"
                    },
                    "expiration_date": {
                        "type": "string",
                        "example": "2023-06-06 10:00:00"
                    },
                    "meta_data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "type": "string",
                        "example": "0741236587"
                    },
                    "uuid": {
                        "type": "string",
                        "example": "9e4cdbc3-83ee-4f5f-820a-4791a1c804fa"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "George"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Puscas"
                    },
                    "email": {
                        "type": "string",
                        "example": "email@email.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_Auth": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}